+++ webroot/js/src/ccfparser.js
@@ -1,4 +1,37 @@
+// import {server, ccfOptions} from "./config/ccf_config_nl";
 "use strict";
+// ugly for now, d
+var server = 'http://www.huc.localhost/clarin_cmdi_forms/';
+var ccfOptions = {
+    uploadButton: {
+        actionURI: server + 'upload.php'
+    },
+    submitButton: {
+        actionURI: 'create_record.php',
+        label: 'Submit'
+    },
+    saveButton: {
+        actionURI: 'create_record.php',
+        label: 'Save'
+    },
+    resetButton: {
+        actionURI: 'javascript:history.back()',
+        label: 'Back'
+    },
+    language: 'en',
+    alert: {
+        mandatory_field: 'This field is mandatory!',
+        mandatory_field_box: ' : mandatory!',
+        no_valid_date: 'This is not a valid date!',
+        no_valid_date_box: ': not a valid date!',
+        date_string: 'yyyy-mm-dd',
+        int_field: 'The value of this field must be an integer!',
+        int_field_box: 'must be an integer!',
+        attr_not_empty_field: 'Attribute(s) mandatory'
+    }
+};
+// import a module for side-effects only # d
+// not recommended practice, some modules set up some global state that can be used by other modules.
 var serverurl = "http://localhost:8888/server.php";
 fetch(serverurl).then(function (response) {
     response.json().then(function (json) {
@@ -8,7 +41,7 @@ fetch(serverurl).then(function (response) {
         });
     });
 });
-console.log('hello typescript');
+console.log('hello typescriasdasdffpt');
 var objectDisplay = true;
 var objectLevel = 1;
 var validationProfiles = {};
@@ -850,7 +883,7 @@ function duplicateField(obj, set) {
 }
 function duplicateComponent(obj, set) {
     var next = clone.nextClonePostfix();
-    clonedComponent = $(set).find("div[data-name='" + obj.name + "']").first().clone();
+    var clonedComponent = $(set).find("div[data-name='" + obj.name + "']").first().clone();
     clonedComponent.addClass("clonedComponent");
     clonedComponent.attr("id", clonedComponent.attr("id") + '_' + next);
     clonedComponent.find(".compBtn").each(function () {
@@ -887,8 +920,7 @@ function duplicateComponent(obj, set) {
     //       console.log(target);
     //       $(this).attr('target', target + '_' + next);
     //    });
-    var clonedComponent, find;
-    (".optionalCompBtn").each(function () {
+    clonedComponent.find(".optionalCompBtn").each(function () {
         $(this).attr('value', "✗");
         $(this).on("click", showComponentFields);
     });