+++ webroot/js/src/ccfparser.ts
@@ -1,4 +1,43 @@
+// import {server, ccfOptions} from "./config/ccf_config_nl";
+
 "use strict";
+
+// ugly for now, d
+let server = 'http://www.huc.localhost/clarin_cmdi_forms/';
+
+let ccfOptions = {
+    uploadButton: {
+        actionURI: server + 'upload.php'
+    },
+    submitButton: {
+        actionURI: 'create_record.php',
+        label: 'Submit'
+    },
+    saveButton: {
+        actionURI: 'create_record.php',
+        label: 'Save'  
+    },
+    resetButton: {
+        actionURI: 'javascript:history.back()',
+        label: 'Back'
+    },
+    language: 'en',
+    alert: {
+      mandatory_field: 'This field is mandatory!',
+      mandatory_field_box: ' : mandatory!',
+      no_valid_date: 'This is not a valid date!',
+      no_valid_date_box: ': not a valid date!',
+      date_string: 'yyyy-mm-dd',
+      int_field: 'The value of this field must be an integer!',
+      int_field_box: 'must be an integer!',
+      attr_not_empty_field: 'Attribute(s) mandatory'
+    }
+};
+
+
+
+// import a module for side-effects only # d
+// not recommended practice, some modules set up some global state that can be used by other modules.
 
 let serverurl = "http://localhost:8888/server.php";
 fetch(serverurl).then(function (response) {
@@ -12,7 +51,8 @@ fetch(serverurl).then(function (response) {
 });
 
 
-console.log('hello typescript');
+
+console.log('hello typescriasdasdffpt');
 
 let objectDisplay = true;
 let objectLevel = 1;
@@ -887,7 +927,7 @@ function duplicateField(obj, set) {
 
 function duplicateComponent(obj, set) {
     var next = clone.nextClonePostfix();
-    clonedComponent = $(set).find("div[data-name='" + obj.name + "']").first().clone();
+    let clonedComponent = $(set).find("div[data-name='" + obj.name + "']").first().clone();
     clonedComponent.addClass("clonedComponent");
     clonedComponent.attr("id", clonedComponent.attr("id") + '_' + next);
     clonedComponent.find(".compBtn").each(
@@ -928,7 +968,7 @@ function duplicateComponent(obj, set) {
     //       console.log(target);
     //       $(this).attr('target', target + '_' + next);
     //    });
-    let clonedComponent.find(".optionalCompBtn").each(function () {
+    clonedComponent.find(".optionalCompBtn").each(function () {
         $(this).attr('value', "✗");
         $(this).on("click", showComponentFields);
     });