+++ webroot/js/src/config/ccf_config_en.js
@@ -1,5 +1,6 @@
-var server = 'http://www.huc.localhost/clarin_cmdi_forms/';
-var ccfOptions = {
+"use strict";
+server = 'http://www.huc.localhost/clarin_cmdi_forms/';
+ccfOptions = {
     uploadButton: {
         actionURI: server + 'upload.php'
     },