+++ webroot/js/src/config/ccf_config_nl.js
@@ -1,5 +1,8 @@
-var server = 'http://www.huc.localhost/clarin_cmdi_forms/';
-var ccfOptions = {
+"use strict";
+// variables server and ccfOptions undeclared out for the problems, stupid editor thing, the files are not included...
+server = 'http://www.huc.localhost/clarin_cmdi_forms/';
+console.log('modules');
+ccfOptions = {
     uploadButton: {
         actionURI: server + 'upload.php'
     },
@@ -27,3 +30,5 @@ var ccfOptions = {
         attr_not_empty_field: 'Attribuut: $attributename verplicht'
     }
 };
+// export { server, ccfOptions };
+// https://stackoverflow.com/questions/41993445/exporting-a-function-on-typescript-declaration-or-statement-expected